Document process of installing go version 1.12
go-orbit-db only works with version 1.12.  For users that are on 1.13
but don't want to roll back system-wide, we can use "go get" to
install an old version of golang.  This is the easiest way to
installing multiple go versions and teaching your system to let them
co-exist.

This patch documents the two needed commands, notes when they are
needed, and informs the user how to use the resulting version 1.12
binary.